requirements.txt
# Core dependencies - Tested with Python 3.12.5
qdrant-client>=1.14.2,<2.0.0
sentence-transformers>=4.1.0,<5.0.0
langchain>=0.3.25,<0.4.0
langchain-community>=0.3.24,<0.4.0  # For BM25Retriever and other community components
langchain-text-splitters>=0.3.8,<0.4.0  # For RecursiveCharacterTextSplitter
rank-bm25>=0.2.2,<1.0.0  # BM25 keyword search algorithm
tiktoken>=0.7.0,<1.0.0
python-dotenv>=1.1.0,<2.0.0
watchdog>=6.0.0,<7.0.0
pydantic>=2.8.2,<3.0.0
fastapi>=0.115.12,<1.0.0
uvicorn>=0.34.2,<1.0.0
httpx>=0.27.0,<1.0.0
pyyaml>=6.0.1,<7.0.0
requests>=2.32.3,<3.0.0  # Used in scripts
psutil>=5.9.0,<6.0.0  # System and process monitoring

# MCP framework
mcp>=1.8.1,<2.0.0  # Model Context Protocol SDK

# GitHub integration (v0.3.0)
pygithub>=2.6.1,<3.0.0  # GitHub API client
gitpython>=3.1.44,<4.0.0  # Git operations

# Optional performance improvements for Python 3.12
orjson>=3.10.18,<4.0.0  # Faster JSON parsing
numpy>=2.1.0,<3.0.0   # NumPy 2.x for Python 3.12
torch>=2.4.0,<3.0.0   # If using GPU acceleration